--- app/src/main/java/com/challengeonair/challengeonairandroid/model/api/response/UserProfileResponse.kt
@@ -1,25 +1,0 @@
-package com.okaka.challengeonairandroid.model.api.response
-
-import com.google.gson.annotations.SerializedName
-
-data class UserProfileResponse(
-    @SerializedName("user_nick_name") val userNickName: String,
-    @SerializedName("user_body") val userBody: String,
-    @SerializedName("image_url") val imageUrl: String,
-    @SerializedName("point") val point: Int
-)
-
-data class UserProfileUpdateRequest(
-    @SerializedName("user_nick_name") val userNickName: String? = null,
-    @SerializedName("user_body") val userBody: String? = null,
-    @SerializedName("image_url") val imageUrl: String? = null
-)
-
-data class UserProfileUpdateResponse(
-    @SerializedName("user_id") val userId: String
-)
-
-data class UserProfileSpecificResponse(
-    @SerializedName("user_nick_name") val userNickName: String,
-    @SerializedName("image_url") val imageUrl: String
-)
