+++ app/src/main/java/com/okaka/challengeonairandroid/model/api/service/UserProfileApi.kt
@@ -2,6 +2,7 @@ package com.okaka.challengeonairandroid.model.api.service
 
 import com.okaka.challengeonairandroid.model.api.response.ApiResponse
 import com.okaka.challengeonairandroid.model.api.response.UserProfileResponse
+import com.okaka.challengeonairandroid.model.api.response.UserProfileSpecificResponse
 import com.okaka.challengeonairandroid.model.api.response.UserProfileUpdateRequest
 import com.okaka.challengeonairandroid.model.api.response.UserProfileUpdateResponse
 import retrofit2.http.Body
